+++ tomcat/src/main/java/org/apache/catalina/controller/Controller.java
@@ -1,7 +1,7 @@
 package org.apache.catalina.controller;
 
-import org.apache.coyote.http11.message.HttpRequest;
-import org.apache.coyote.http11.message.HttpResponse;
+import org.apache.coyote.http11.message.request.HttpRequest;
+import org.apache.coyote.http11.message.response.HttpResponse;
 
 public interface Controller {
     void service(HttpRequest request, HttpResponse response) throws Exception;
