+++ tomcat/src/main/java/org/apache/catalina/controller/HomeController.java
@@ -1,12 +1,12 @@
 package org.apache.catalina.controller;
 
 import java.nio.charset.StandardCharsets;
-import org.apache.coyote.http11.message.HttpRequest;
-import org.apache.coyote.http11.message.HttpResponse;
-import org.apache.coyote.http11.message.HttpResponseHeader;
+import org.apache.coyote.http11.message.request.HttpRequest;
+import org.apache.coyote.http11.message.response.HttpResponse;
+import org.apache.coyote.http11.message.response.HttpResponseHeader;
 import org.apache.coyote.http11.message.HttpStatus;
 import org.apache.coyote.http11.message.StaticResourceExtension;
-import org.apache.coyote.http11.message.StatusLine;
+import org.apache.coyote.http11.message.response.StatusLine;
 
 public class HomeController implements Controller {
