+++ tomcat/src/main/java/org/apache/catalina/controller/RegisterController.java
@@ -7,11 +7,11 @@ import java.util.HashMap;
 import java.util.Map;
 import org.apache.catalina.cookie.ResponseCookie;
 import org.apache.catalina.session.SessionManager;
-import org.apache.coyote.http11.message.HttpRequest;
-import org.apache.coyote.http11.message.HttpResponse;
-import org.apache.coyote.http11.message.HttpResponseHeader;
+import org.apache.coyote.http11.message.request.HttpRequest;
+import org.apache.coyote.http11.message.response.HttpResponse;
+import org.apache.coyote.http11.message.response.HttpResponseHeader;
 import org.apache.coyote.http11.message.HttpStatus;
-import org.apache.coyote.http11.message.StatusLine;
+import org.apache.coyote.http11.message.response.StatusLine;
 
 public class RegisterController implements Controller {
