+++ tomcat/src/main/java/org/apache/coyote/http11/message/response/HttpResponse.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.message;
+package org.apache.coyote.http11.message.response;
 
 public class HttpResponse {
